Consolidate JS Code to app.js file and delete JS modules
