+++ src/index.html
@@ -88,7 +88,7 @@
         </svg>
       </div>
     </header>
-    <div id="page" class="about">
+    <div id="page" class="homePage">
       <section class="hero">  
         <article class="card">
           <h3>"Say It Loud!"</h3>
