--- src/models/db.js
@@ -1,19 +1,0 @@
-const artistData = (fullName, songs, blurb, imgUrl) => ({
-  fullName,
-  songs,
-  blurb,
-  imgUrl,
-});
-
-const db = {
-  artists: [
-    artistData('Ivy Queen', ['Quiero Bailar', 'CELEBRANDO', 'Quiero Saber'], 'Ivy Queen is Reggaeton royalty, her empowering lyrics and electrifying beats have left an indelible mark on the music scene. A true trailblazer inspiring a global wave of fans.', 'https://i.pinimg.com/474x/8a/b0/5d/8ab05d648464d4023039ff23fe988998.jpg'),
-    artistData('Ivy Queen', ['Quiero Bailar', 'CELEBRANDO', 'Quiero Saber'], 'Ivy Queen is Reggaeton royalty, her empowering lyrics and electrifying beats have left an indelible mark on the music scene. A true trailblazer inspiring a global wave of fans.', 'https://i.pinimg.com/474x/8a/b0/5d/8ab05d648464d4023039ff23fe988998.jpg'),
-    artistData('Ivy Queen', ['Quiero Bailar', 'CELEBRANDO', 'Quiero Saber'], 'Ivy Queen is Reggaeton royalty, her empowering lyrics and electrifying beats have left an indelible mark on the music scene. A true trailblazer inspiring a global wave of fans.', 'https://i.pinimg.com/474x/8a/b0/5d/8ab05d648464d4023039ff23fe988998.jpg'),
-    artistData('Ivy Queen', ['Quiero Bailar', 'CELEBRANDO', 'Quiero Saber'], 'Ivy Queen is Reggaeton royalty, her empowering lyrics and electrifying beats have left an indelible mark on the music scene. A true trailblazer inspiring a global wave of fans.', 'https://i.pinimg.com/474x/8a/b0/5d/8ab05d648464d4023039ff23fe988998.jpg'),
-    artistData('Ivy Queen', ['Quiero Bailar', 'CELEBRANDO', 'Quiero Saber'], 'Ivy Queen is Reggaeton royalty, her empowering lyrics and electrifying beats have left an indelible mark on the music scene. A true trailblazer inspiring a global wave of fans.', 'https://i.pinimg.com/474x/8a/b0/5d/8ab05d648464d4023039ff23fe988998.jpg'),
-    artistData('Ivy Queen', ['Quiero Bailar', 'CELEBRANDO', 'Quiero Saber'], 'Ivy Queen is Reggaeton royalty, her empowering lyrics and electrifying beats have left an indelible mark on the music scene. A true trailblazer inspiring a global wave of fans.', 'https://i.pinimg.com/474x/8a/b0/5d/8ab05d648464d4023039ff23fe988998.jpg'),
-  ],
-};
-
-export default db;
