--- src/views/components/button.js
@@ -1,3 +1,0 @@
-const button = (className, text) => `<button type="button" class="${className}">${text}</button>`;
-
-export default button;
