--- src/views/components/heading.js
@@ -1,2 +1,0 @@
-const header = (type, text, className) => `<h${type} class="${className}">${text}</h${type}>`;
-export default header;
