--- src/views/components/logo.js
@@ -1,6 +1,0 @@
-const logo = (text) => `
-    <a href="#home">
-      <figure class="logoBox"><div class="logoText">${text}</div></figure>
-    </a>
-  `;
-export default logo;
