--- src/views/pages/about.js
@@ -1,7 +1,0 @@
-const about = () => `
-    <div class="page">
-      about page
-    </div>
-  `;
-
-export default about;
