--- src/views/pages/home.js
@@ -1,5 +1,0 @@
-const home = () => `
-      home page
-  `;
-
-export default home;
